requirements.txt
absl-py==2.3.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.9
dm_control==1.0.34
etils==1.13.0
filelock==3.20.1
fonttools==4.61.1
fsspec==2025.12.0
glfw==2.10.0
gym==0.26.2
gym-notices==0.1.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
labmaze==1.0.6
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mujoco==3.4.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
PyOpenGL==3.1.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sympy==1.14.0
termcolor==3.2.0
-e git+https://github.com/neuromatch/tonic.git@59dedfa3b0988b23446868916d9149b21f32fbf9#egg=tonic
torch==2.9.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wrapt==2.0.1
zipp==3.23.0
